# Cache field encoders and setter