Before:

```
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    2000	    731466 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      20	  68835818 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.070s
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    2000	    728904 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      20	  68418671 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.047s
```

After:

```
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    2000	    515373 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      20	  55145695 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	2.292s
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    2000	    536994 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      20	  52838184 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	2.287s
```