@alex 2019-08-23
Got a bit faster even:

```
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    3000	    445487 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      30	  41887096 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.593s
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    3000	    443848 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      30	  41074566 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	2.693s
```